requirements.txt
huggingface-hub
numpy
pandas
python-dotenv
requests
torch
lmdb
colorlog
tensorboardX
colorama
torch-geometric